TEST: Melhora a config do projeto Selenium (#184)
* test: modifica o arquivo de build do gradle para as novas configs

* test: remove as properties e centraliza as configs no env

* test: move o basetest de lugar tornando uma classe reutilizavel

* test: adiciona o changeset

* test: adiciona algumas sugestoes do code rabbit para maior legibilidade do codigo
