+++ .changeset/pretty-snakes-fix.md
@@ -1,0 +1,5 @@
+---
+"selenium-ui-tests": patch
+---
+
+test: melhora a configuração do projeto Selenium
