--- tests/selenium/src/test/resources/application.properties
@@ -1,42 +1,0 @@
-# Configurações do Browser
-browser=chrome
-headless=false
-
-# Configurações de URL
-baseUrl=http://localhost:3000
-apiUrl=http://localhost:3000/api
-
-# Configurações de Timeout
-implicitWait=0
-explicitWait=0
-pageLoadTimeout=30
-scriptTimeout=30
-
-# Configurações de Screenshot
-screenshotOnFailure=true
-screenshotOnSuccess=false
-screenshotPath=build/screenshots
-
-# Configurações de Log
-logLevel=INFO
-logPath=build/logs
-
-# Configurações de Teste
-retryCount=1
-parallelExecution=true
-maxThreads=4
-
-# Configurações de Ambiente
-environment=local
-testDataPath=src/test/resources/testdata
-
-# Configurações de Credenciais (para testes que precisam de login)
-username=${E2E_USERNAME}
-password=${E2E_PASSWORD}
-
-# Configurações de Proxy (se necessário)
-proxyEnabled=false
-proxyHost=
-proxyPort=
-proxyUsername=
-proxyPassword=
